[Feat] 데이터 버전을 올려 이전 데이터에 없던 인수인 HealLevel을 추가
+ 적의 방어력 추가
+ 적의 체력 증가
+ 강화수치 수정
+ HealLevel에 따른 회복

+++ Assets/03.Scripts/Enhancement/ResultWindow.cs
@@ -32,7 +32,7 @@ public class ResultWindow : MonoBehaviour
             return; 
         }
 
-        int random = Random.Range(0, 100);
+        float random = Random.Range(0f, 100f);
 
         GameManager.Instance.ChangeMoney(-PlayerEquip.Instance.EquipmentSlot[PlayerEquip.Instance.CheckEquipNumber].PriceItem());
